LibGPU+LibSoftGPU: Move RasterizerOptions into LibGPU
